Baget_2024 version 1.8.3 (For Order Form)

+++ src/main/java/com/example/baget/parts/ProfilListDTO.java
@@ -12,7 +12,7 @@ public class ProfilListDTO {
     @Size(max = 30)
     private String description;
 
-    private Double profilWidth;
+    private Long profilWidth;
 
     private Double onHand;
